--- src/ensembl/src/shared/components/hoverbox/index.ts
@@ -1,2 +1,0 @@
-export { default as Hoverbox } from './Hoverbox';
-export { default as HoverboxExpandableContent } from './hoverbox-expandable-content/HoverboxExpandableContent';
