+++ src/ensembl/src/shared/components/toolbox/index.ts
@@ -1,0 +1,2 @@
+export { default as Toolbox } from './Toolbox';
+export { default as ToolboxExpandableContent } from './toolbox-expandable-content/ToolboxExpandableContent';
