+++ src/ensembl/src/shared/components/toolbox/toolbox-expandable-content/ToolboxExpandableContent.scss
