+++ src/ensembl/stories/shared-components/toolbox/Toolbox.stories.scss
